DX-1080: Use mixin pattern for x-element.
This pattern is copied from how polymer-element works:

https://github.com/Polymer/polymer/blob/master/polymer-element.js

The mixin pattern is detailed here:

http://justinfagnani.com/2015/12/21/real-mixins-with-javascript-classes/
